Clipping complete! Kinda
Compile the code
Then run it multiple times

Sometimes, it renders the entire model instead of rendering clipped models.
It seems to do this for one or more planes at random.

The planes still clip and add triangles to the list, but the triangle list stays what it was originally for some reason